Tidied up some more examples, removing the redundant use of <j:expr value=, when just ${foo} will do now.
Added a couple of Ant Project helper methods to the AntTagLibrary, to create Ant projects in a single place in the codebase.

Also made the Werkz tag library use lazy construction for finding goals, so that a project.jelly can define pre-conditions/pre/post callbacks, then
import a maven.jelly file to actually define all the goals and run the build.

The 'demo.maven' target shows this in action, allowing a local project.jelly to just define the callbacks then invoke the main maven.jelly


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk/jelly-tags/ant@134967 13f79535-47bb-0310-9956-ffa450edef68