First cut of the http tag library using commons-httpclient to script HTTP operations inside Jelly. Its not quite working yet but its close.
When this library is working properly we should be able to use this to perform HTTP based SOAP scripting and unit testing.
For example we should be able to do
<babelfish:translate from="en" to="fr">Hello there my friend</babelfish:translate>

and for the tag to do all the SOAP and HTTP malarkey inside the tag definition, which could be written in Jelly script itself.

Similarly we could build macros for testing web applications. Maybe iterating over test data in XML documents to test that web applications perform properly.

e.g.

<x:forEach select="$tests/testcase">
	<x:set var="url" select="string(url)"/>
	<test:httpGet url="${url}" expects="100"/>

etc.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@134999 13f79535-47bb-0310-9956-ffa450edef68