Added support for dynamic jelly beans. These are similar to <taskdef> in Ant but in this case any Java Bean can be bound to a Jelly tag then any 'invoke' method can be called.
So for example any Runnable object can be dynamically bound into a dynamic tag library and invoked as a tag.
The method could be run(), invoke(), execute() or whatever. 
The Jexl and tag attributes can set all the properties of the bean.

So dynamic tag libraries can now be defined using regular beans or via jelly script.
One more dynamic tag that uses a static method might be interesting to do...


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135012 13f79535-47bb-0310-9956-ffa450edef68