Patched the Ant Tag adapter to follow more closely the task lifecycle document, kindly outlined by Stefan. This new AntTag should replace the old stuff: TaskTag, OtherAntTag, DataTypeTag and AntTagScript. We should delete those once we've checked that AntTag works properly.
There does appear to be still a problem with references; this checkin is hopefully so that bob can take a once-over and spot any schoolboy errors.
I think the logic is fine for Tasks though we may need some more work on nested properties/elements and data types.
We're nealy there now...


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135023 13f79535-47bb-0310-9956-ffa450edef68