Added a "dist:install" goal to install Jelly on the user's system. This creates a ${maven.dist.install.dir} dir, and copies all the files needed by Jelly to run (which are generated in "dist:build"'s preGoal.
Though, I can't generate the JELLY_HOME and the PATH, so I added a
big warning telling the user that he needs to set those variables.

Note : I defaulted the ${maven.dist.install.dir} to /usr/local/jelly,
which is a Linux path. I should add a system detection, so that
both worlds are happy :)


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135079 13f79535-47bb-0310-9956-ffa450edef68