Enhanced the JUnit tag library so that TestSuite and TestCase objects can be created in Jelly script so that unit testing becomes easy to do via Jelly scripts. This also means that Jelly can be integrated easily into existing test runners.
See the src/test/org/apache/commons/jelly/junit/ directory for details of how this works; there you'll find a suite.jelly which is the test suite and a TestJUnit.java class which is a simple class to bind the suite.jelly into Ant's default JUnit test runner.
This class should be usable in any JUnit test runner, since it provides a class with a static suite() method, yet internally it uses one or more Jelly scripts to actually create the individual TestCase objects.

So the new JellyUnit capabilities can be seamlessly integrated with traditional Java test cases. So users can now use either Java or Jelly; intermixing the two to use the best tool for the job. So now Jelly could be used for XML, SOAP, SQL, HTTP & JMS testing and such like.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135087 13f79535-47bb-0310-9956-ffa450edef68