Patched the Script implementation mechanism for 2 reasons
(i) to try make a single Script instance cacheable and multi-threaded so that in server side deployments, the scripts can be cached and reused across multiple threads to boost performance and lower RAM usage

(ii) to introduce the TagFactory so that a tag can be bound to a Tag instance at runtime, rather than just at parse time.
This makes many things much simpler, particularly in the area of dynamic tags.

For example the Jeez library unifies the define tag library with the ant tag library, both of which can create dynamic tags, at run time.
Until this change was done, it was not possible to return a single efficient Tag implementation at parse time. So now at parse time a TagLibrary can create a TagFactory which at runtime decides which Tag instance to use.
This chance helps the multi-threaded nature also, so that different threads can create their own tag implemention.


Right now the current policy is for a Script object to be reusable across threads (which needs testing), then each thread will create & cache 1 Java object per tag on the page. Each Tag is cached in ThreadLocal storage. If a Script goes out of scope then all this should be GC'd.
So we've a neat, simple,  multi-threaded, pooled solution that should cause minimum object allocation yet allow flexible bean style coding and not require byte code generation (and all the classloading / GC issues that entails).

Also for the Maven users, the <tagdef> is now working and can be used inside Maven now, which simplifies things greatly - no need to create a new tag library & new tag, just use <tagdef> and add a new tag dynamically to the maven (Jeez) tag library.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135120 13f79535-47bb-0310-9956-ffa450edef68