Patched to use Ant 1.5. Also removed the local copy of Ant's IntrospectionHelper that was added as a temporary hack.
Added a patch for Ant tags which can be both a Task and a nested object. 
For example the <manifest> tag can be a ManifestTask if used by itself, of if it is embedded inside a <jar> Task then the Jar.addConfiguredManifest() is called instead.
This fixes the manifest bug that was introduced by the move of Maven to Ant 1.5.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk/jelly-tags/ant@135155 13f79535-47bb-0310-9956-ffa450edef68