Added support for evaluateAsValue() which (though I don't like the name) will evaluate any nested expressions.
The evaluateAsString(), evaluateAsIterator() and evaluateAsBoolean() methods now use this new method.

The reason for this is that sometimes an expression may evaluate to be an Expression object. So the evaluateAsValue() will evaluate the expression, then if the result is another Expression it will keep evaluating it, recursively until its not an Expression.

Care should be taken when making infinite Expression nesting!

The reason for this fix is that often Maven will populate the Maven JellyContext with variables which are Expressions (rather than the values of the expressions). So that ${maven.foo} when evaluated would lookup the value of 'maven.foo', which could return an Expression.
This change will hopefully provide auto-unwrapping of these 'nested expressions'


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135215 13f79535-47bb-0310-9956-ffa450edef68