Added a <j:useBean> tag to both mimick the similar JSP tag but also to combine it with the <j:setProperties> tag so that a fully configured bean can be created easily by a single tag.
<j:useBean var="foo" class="com.acme.MyCustomer" name="James" location="London" role="Hacker"/>

Where the above will create a new variable called 'foo' of type com.acme.MyCustomer and then will set the name, location and role properties to the values given.

This tag is also useful as a base class for constructing arbitrary objects of certain kinds that can be configured and used in some script - such as to plug in arbitrary Actions into a workflow style script or in a simple testing framework etc.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135224 13f79535-47bb-0310-9956-ffa450edef68