Applied Robert Leftwich's excellent patches to add full support for <x:transform> for executing XSLT on some XML. The XML or the XSLT can come from the tags body just like the <x:transform> tag in JSTL.
Ultimately this tag can be used to implement sophisticated XSLT pipelining such as...

<x:transform xslt="a.xsl">
    <x:transform xslt="b.xsl">
        <foo>
            <bar name="${some.expression}"/>
        </foo>
    </x:transform>
</x:transform>

which can then be used in association with servlet engines, SOAP services or just generating static content.

It'd be nice to optimize the implementation at some point so that it can work totally on SAX in the above example. But this is an awesome start.

Also do check out the groovy unit test cases such as the Schematron examples. It'd be nice to maybe generate a native Jelly implementaiton of Schematron...


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135229 13f79535-47bb-0310-9956-ffa450edef68