Some JellySwing changes.
Firstly added support for <borderLayout> and <borderAlign align="center"> style of tags to use the Swing BorderLayout.
Added support for <vbox> and <hbox> to create Swing Box panels (which have their own inbuilt BoxLayout)

Also added colfill="true" and rowfill="true" options on <td> so that columns can stretch

Finally I've improved the JellyRunner so that its a little dialog, with a history field and a button so that reevaluating the same script can be much easier. Just hit save in your editor and hit the button and the Jelly script executes immediately - no more waiting for a JVM to start.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135234 13f79535-47bb-0310-9956-ffa450edef68