Minor patch so that DynaTag objects can declare the types of their dynamic properties to facilitate better expression type coercion.
For example a DynaTag might have an attribute which is of type Expression which the DynaTag wishes to evaluate however it wishes without it just being turned into an Object.
So this method allows DynaTag objects to return, for example, that an attribute is of type Expression.class allowing them to coerce the Expression themselves.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135252 13f79535-47bb-0310-9956-ffa450edef68