Added support in JellySwing for simple TableModels which use any pluggable Jelly Expressions for defining the columns.
This allows any List of objects to be easily rendered in a table in JellySwing without any programming involved. Its hoped this kind of thing could be ported to SWT too!

A simple example has been added which will display all the variables in the current context in a table.
This could be improved somewhat, but shows that a List of property names can be displayed in a table, showing the value, type, row index and the name of the variable; all powered by a List of Strings.

Another good demo could be to render a table of beans (such as some Maven POM objects) in a table.
Or to use commons-beanutils or commons-sql to get some DynaBeans and render those in a table using expressions.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135255 13f79535-47bb-0310-9956-ffa450edef68