Added SOAP support via Apache Axis.
This is based on a patch submitted by Jim Birchfield which allows SOAP services to be invoked using Apache Axis and the results output.

Its early days and we'll improve this support considerably over time but at least the basics are working for now.

There's a demo invokable via

  maven demo:soap

which invokes a sample Axis echo service and an xmethods.net service.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135289 13f79535-47bb-0310-9956-ffa450edef68