Committed patch supplied by Robert Leftwich which creates a Jetty tag library for creating HTTP servers for testing web requests using an embedded Jetty instance.
It could be that there is some crossover with 

Latka
http://jakarta.apache.org/commons/latka/index.html

Anteater
http://aft.sourceforge.net/

or Cactus
http://jakarta.apache.org/cactus/

I'm not personally expert enough to know what should be done in this area. So I'm committing this and then the discussion between the various groups can start. 
It might be that one of the above projects could be a better home for the above code - lets see what the community at large thinks.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135342 13f79535-47bb-0310-9956-ffa450edef68