Added support for NamespaceAwareTags. This allows Tag implementations to be given the namespace context in which they are used. This allows tags to, for example, map namespace qualified names to namespace URIs using namespace prefixes.
For example in XSD or SOAP or BPEL styles of languages, qualified names like "x:foo" can be used. The 'x' is a prefix. The namespace context maps prefixes 'x' to the URI.

This patch also means that all the XPath expressions in the XML library, JSL and JellyUnit can also make use of the XPath prefixes that are available.
This allows full namespace-based XPath to be used.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135357 13f79535-47bb-0310-9956-ffa450edef68