Patch to fix dIon's bug with JSL. The fix was for TagScript to properly set the parent tag each time its used; there was a danger of the Tag being cached and its old parent being used when inside loops. This should now fix any similar related bugs.
Moving forward, it might be worth optimizing the TagScript code so that the Script.run() method takes the parent Tag to use. Then the thread local tag hierarchy can be maintained on the stack, rather than using ThreadLocal variables that we currently use. This would be faster and simpler I suspect.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135384 13f79535-47bb-0310-9956-ffa450edef68