Patch applied for the bug found by dIon. Also added a few more test cases to fully test this feature to ensure it doesn't break again.
The bug essentially was that the <applyTemplates> is often invoked from a different XMLOutput, which needs to be associated with the Stylesheet so that as new <template>s fire they use the correct XMLOutput.
This is required as things like <j:set> use a different XMLOutput to buffer up its body.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135387 13f79535-47bb-0310-9956-ffa450edef68