I've enabled the JellyUnit test case that Peter wrote and added another to both test for a regular nested <filterset> and one using a refid.
Also added a fix for it too! The short answer is the problem was due to treating <filter> as a new task rather than a nested property of <filterset>

Again this is due to the complexity of the Ant behaviour. Sometimes Ant tags are meant to be Tasks, sometimes they are meant to be nested properties, and this can get confusing. I think the correct logic right now is that...

  a tag should be considered task if it is the root node in the tree of Ant tags and the name of the tag is a valid Ant task name, otherwise treat it as a nested property or datatype.

So the root tag can be a <filterset id="...">. Also there are times when a task shouldn't be treated as a task but instead as a nested property. e.g. <filter> and <manifest> are tasks but should be treated as nested properties.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135389 13f79535-47bb-0310-9956-ffa450edef68