Patch to allow the HTML tag library full control over NeckoHTML's ability to convert elements or attributes to upper or lowercase.
You can now specify if you wish case conversion to be applied when parsing HTML. 

e.g. to upper case the elements but lowercase the attributes you can do...

    <html:parse var="doc" html="index.html" element="upper" attribute="lower"/>

or to make everything lower case

    <html:parse var="doc" html="index.html" element="lower" attribute="lower"/>


The default is to leave case as it was but to match element case (so that open and close elements have the same case).


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/jelly/trunk@135450 13f79535-47bb-0310-9956-ffa450edef68